requirements.txt
# Core ML Libraries
scikit-learn>=1.5.0
numpy>=2.0.0
pandas>=2.0.0

# MLflow for experiment tracking
mlflow>=2.15.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0.0

# Pipelines
zenml>=0.64.0

# Optimization
optuna>=3.6.0

# Testing
pytest>=7.4.0